Add optional `vocational_class` field to student forms and views
This update makes the `vocational_class` field optional in the `VocationalStatus` model with support for `null` and `blank`. Adjustments were made in forms, views, and templates to handle conditional display and management of this field, specifically for schools with the abbreviation "BMA". Migration file included to apply the model change.